# tests/golden/snapshots/openapi.json
{
  "components": {
    "schemas": {
...
      "Committed": {
        "description": "What a commit wrote.\n\nAttributes:\n    commit_id: The `commit_log` row.\n    applied_operation_ids: Applied operations, in order.\n    entity_ids: For each applied `create_entity` seq, the entity it resolved to.\n    index_pending: Entity ids the similarity index could not embed yet (the\n        embedder was unavailable). The commit still happened; only the\n        search index is behind — `reindex_missing` catches these up later.",
        "properties": {
          "applied_operation_ids": {
            "items": {
              "type": "integer"
            },
            "title": "Applied Operation Ids",
            "type": "array"
          },
          "commit_id": {
            "title": "Commit Id",
            "type": "integer"
          },
          "entity_ids": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Entity Ids",
            "type": "object"
          },
          "index_pending": {
            "default": [],
            "items": {
              "type": "integer"
            },
            "title": "Index Pending",
            "type": "array"
          }
        },
        "required": [
          "commit_id",
          "applied_operation_ids",
          "entity_ids"
        ],
        "title": "Committed",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JsonValue": {},
...
      "OperationRecord": {
        "description": "One operation of a proposal, with its review state.",
        "properties": {
          "classification": {
            "title": "Classification",
            "type": "string"
          },
          "edited_payload": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Edited Payload"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "op_type": {
            "title": "Op Type",
            "type": "string"
          },
          "payload": {
            "additionalProperties": true,
            "title": "Payload",
            "type": "object"
          },
          "rationale": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rationale"
          },
          "seq": {
            "title": "Seq",
            "type": "integer"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "target_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Id"
          },
          "target_kind": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Kind"
          }
        },
        "required": [
          "id",
          "seq",
          "op_type",
          "payload",
          "classification",
          "target_kind",
          "target_id",
          "status",
          "edited_payload",
          "rationale"
        ],
        "title": "OperationRecord",
        "type": "object"
      },
      "ProposalRecord": {
        "description": "A proposal and its operations in `seq` order.",
        "properties": {
          "applied_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Applied At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "operations": {
            "items": {
              "$ref": "#/components/schemas/OperationRecord"
            },
            "title": "Operations",
            "type": "array"
          },
          "origin": {
            "title": "Origin",
            "type": "string"
          },
          "source_id": {
            "title": "Source Id",
            "type": "integer"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          }
        },
        "required": [
          "id",
          "origin",
          "source_id",
          "status",
          "summary",
          "created_at",
          "applied_at",
          "operations"
        ],
        "title": "ProposalRecord",
        "type": "object"
      },
      "ProposalStatus": {
        "description": "Lifecycle of a proposal: open for review, then committed exactly once.",
        "enum": [
          "open",
          "committed"
        ],
        "title": "ProposalStatus",
        "type": "string"
      },
      "ProposalSummary": {
        "description": "A proposal as a row in the review queue, without its operations.\n\nAttributes:\n    id: The proposal id.\n    origin: Which flow produced it (``chat``, ``document``, ...).\n    status: ``open`` or ``committed``.\n    summary: The agent's one-line description.\n    created_at: When it was recorded.\n    applied_at: When it was committed, or None while open.\n    operation_count: How many operations it holds.\n    pending_count: How many of them still await a decision.",
        "properties": {
          "applied_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Applied At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "operation_count": {
            "title": "Operation Count",
            "type": "integer"
          },
          "origin": {
            "title": "Origin",
            "type": "string"
          },
          "pending_count": {
            "title": "Pending Count",
            "type": "integer"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          }
        },
        "required": [
          "id",
          "origin",
          "status",
          "summary",
          "created_at",
          "applied_at",
          "operation_count",
          "pending_count"
        ],
        "title": "ProposalSummary",
        "type": "object"
      },
...
      "Review": {
        "additionalProperties": false,
        "description": "A reviewer's decision on one operation.\n\nAttributes:\n    decision: ``accept``, ``edit`` or ``reject``.\n    edited_payload: The replacement payload, for ``edit`` only.",
        "properties": {
          "decision": {
            "enum": [
              "accept",
              "edit",
              "reject"
            ],
            "title": "Decision",
            "type": "string"
          },
          "edited_payload": {
            "anyOf": [
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/JsonValue"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Edited Payload"
          }
        },
        "required": [
          "decision"
        ],
        "title": "Review",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "CVForge"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/proposals": {
      "get": {
        "description": "List proposals newest first, each with its pending-operation count.\n\nArgs:\n    conn: Read connection.\n    status: ``open`` or ``committed``; omit for all.\n\nReturns:\n    The proposal summaries.",
        "operationId": "list_proposals_api_proposals_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ProposalStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProposalSummary"
                  },
                  "title": "Response List Proposals Api Proposals Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Proposals",
        "tags": [
          "review"
        ]
      }
    },
    "/api/proposals/{proposal_id}": {
      "get": {
        "description": "Return a proposal and its operations.\n\nArgs:\n    conn: Read connection.\n    proposal_id: The proposal id.\n\nReturns:\n    The proposal.\n\nRaises:\n    HTTPException: 404 if the proposal does not exist.",
        "operationId": "get_proposal_api_proposals__proposal_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "proposal_id",
            "required": true,
            "schema": {
              "title": "Proposal Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProposalRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Proposal",
        "tags": [
          "review"
        ]
      }
    },
    "/api/proposals/{proposal_id}/commit": {
      "post": {
        "description": "Apply the accepted and edited operations in one transaction.\n\nNew entities, and existing ones whose name or summary just changed, are\nindexed for similarity search right after (never inside the write\ntransaction — D-D). A failed embedding call never fails the commit: the\naffected ids come back as `index_pending` instead.\n\nArgs:\n    db: The engine.\n    embedder: What indexes new and changed entities; `None` skips indexing.\n    proposal_id: The proposal.\n\nReturns:\n    What was written.\n\nRaises:\n    NotFoundError: Mapped to 404 if the proposal, or a target an operation\n        names, does not exist.\n    ProposalNotOpenError: Mapped to 409 if the proposal was already committed.\n    OperationsPendingError: Mapped to 409 if any operation is still pending.\n    UnsupportedOperationError: Mapped to 422 if an operation cannot be\n        applied by this version (`merge_duplicate`, until M1b).\n    sqlalchemy.exc.IntegrityError: Mapped to 409; the database rejected a\n        row and nothing was written.",
        "operationId": "commit_api_proposals__proposal_id__commit_post",
        "parameters": [
          {
            "in": "path",
            "name": "proposal_id",
            "required": true,
            "schema": {
              "title": "Proposal Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Committed"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Commit",
        "tags": [
          "review"
        ]
      }
    },
    "/api/proposals/{proposal_id}/operations/{operation_id}/review": {
      "post": {
        "description": "Accept, edit or reject one operation; its siblings are untouched.\n\nArgs:\n    db: The engine. The decision is written through `kb.apply`.\n    proposal_id: The proposal the operation must belong to.\n    operation_id: The operation.\n    body: The decision.\n\nReturns:\n    The proposal after the decision.\n\nRaises:\n    HTTPException: 404 if the operation is not part of this proposal.\n    NotFoundError: Mapped to 404 if the operation does not exist.\n    ProposalNotOpenError: Mapped to 409 if the proposal was already committed.\n    InvalidEditError: Mapped to 422 if an edit is missing, changes the\n        operation type, does not validate, or accompanies accept/reject.",
        "operationId": "review_api_proposals__proposal_id__operations__operation_id__review_post",
        "parameters": [
          {
            "in": "path",
            "name": "proposal_id",
            "required": true,
            "schema": {
              "title": "Proposal Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "operation_id",
            "required": true,
            "schema": {
              "title": "Operation Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Review"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProposalRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Review",
        "tags": [
          "review"
        ]
      }
    },
...
  }
}